Remove package references to System.Drawing.Common
This is the last reference we have to System.Drawing.Common.
By removing these we allow folks to select which version of System.Drawing.Common they wish to use
without forcing them to the latest version from 7.0.0 which had breaking changes.  They are free to
stay on 6.0.* if they wish.

This is "safe" because the only purpose of the reference here was for type-forwards for Converter
types we moved between 3.1 and 5.0.  Most folks won't need those forwards, and if they do, they can
simply add a reference to System.Drawing.Common.